apply review for k2-header

+++ runtime-light/k2-platform/k2-api.h
@@ -281,7 +281,7 @@ inline std::expected<k2::SymbolInfo, int32_t> resolve_symbol(void* addr) noexcep
                         .lineno = symbol_info.lineno};
 }
 
-inline size_t code_segment_offset(uint64_t* offset) noexcept {
+inline int32_t code_segment_offset(uint64_t* offset) noexcept {
   return k2_code_segment_offset(offset);
 }
